requirements.txt
python-dotenv
langchain
langchain_openai
azure-functions
pinecone
pypdf
openai
flask